requirements.txt
Django==5.2.1
whitenoise==6.9.0
django-tailwind
django-browser-reload==1.21.0
numpy==1.26.4
Pillow==11.3.0
tensorflow==2.15.0
gunicorn==25.1.0
dotenv
